% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{adequacy.test}
\alias{adequacy.test}
\title{Model adequacy check procedure}
\usage{
adequacy.test(
  fit,
  alpha = 0.05,
  lag.max = min(30, length(fit$residuals) - 1),
  plot = TRUE
)
}
\arguments{
\item{fit}{Fitted ARIMA model(\code{ARIMA} object).}

\item{alpha}{The probability of type 1 error for normality tests. (default = 0.05)}

\item{lag.max}{Maximum length of lag to be tested. (default = \code{min(30, length(fit$residuals) - 1)})}

\item{plot}{If \code{TRUE}, the test result will be presented as a plot.
Otherwise, not. (default = TRUE)}
}
\value{
A list containing the following elements:\tabular{ll}{
  \code{ljung.box} \tab A list of all test results(\code{htest} objects).\cr
  \tab\cr
  \code{plot} \tab \code{ggplot} object of the plot. (Returned only if \code{plot} is TRUE.)\cr
}
}
\description{
Take fitted ARIMA model and perform adequacy checks.
It runs Ljung-Box test with residuals for various lags and print out the test result to the console.
If \code{plot} is TRUE, it also plots the test result.
}
\examples{
# 228 is the station code for SNU
data <- get.subway(228)

# Fit SARIMA model with total on and off passenger counts
SNU.total <- ts(data$total, frequency = 7)
SNU.fit <- auto.arima(SNU.total)

# Normality check
adequacy.test(SNU.fit)
}
\author{
Sanghyun Park, Daun Jeong, and Sehun Kim
}
