% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{detect.change}
\alias{detect.change}
\title{Change point detection}
\usage{
detect.change(data, type = c("mean", "var"), ..., plot = TRUE)
}
\arguments{
\item{data}{Time series data whose change point is to be detected.
The data should be either \code{ts} object or \code{numeric} vector.}

\item{type}{Type of change to be detected.
Should be one of \code{"mean"} and \code{"var"}. (default = "mean")}

\item{...}{Extra arguments for change point detection.
Refer to \code{\link{cpt.mean}} and \code{\link{cpt.var}}.}

\item{plot}{If \code{TRUE}, the chagne points will be plotted.
Otherwise, not. (default = TRUE)}
}
\value{
A list containing the following elements:\tabular{ll}{
  \code{change.point} \tab Change point detection result(\code{cpt} object).\cr
  \tab\cr
  \code{plot} \tab \code{ggplot} object of the plot. (Returned only if \code{plot} is TRUE.)\cr
}
}
\description{
Take time series data and perform change point detection.
It uses \code{\link{cpt.mean}} or \code{\link{cpt.var}} from \code{changepoint} package for change point detection.
If \code{plot} is TRUE, it also plots the point where the changes take place.
}
\examples{
# 228 is the station code for SNU
data <- get.subway(228)

# Change point detection in mean
detect.change(data$total, type = "mean")

# Change point detection in variance
detect.change(data$total, type = "var")
}
\author{
Sanghyun Park, Daun Jeong, and Sehun Kim
}
