% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{get.subway}
\alias{get.subway}
\title{Get subway data}
\usage{
get.subway(
  request,
  group = c("day", "week", "month", "year"),
  FUN = sum,
  ...,
  from = NULL,
  to = NULL,
  split = FALSE
)
}
\arguments{
\item{request}{Station codes for the request.
It should be a \code{numeric} or \code{character} vector indicating the station codes to get.}

\item{group}{Group specification.
Should be one of \code{"day"}, \code{"week"}, \code{"month"}, and \code{"year"}.
Note that \code{"week"} here may not coincide with real world's week concept.
For detail, refer to \code{\link{week}}. (default = "day")}

\item{FUN}{Function to be applied when groupping.
Note that if \code{group} is "day", then this will be ignored. (default = sum)}

\item{...}{Extra arguments for \code{FUN}.}

\item{from}{Character string specifying the start day(inclusive).
It should be a \code{POSIXct} object or a \code{character} which can be casted to \code{POSIXct}.
If \code{NULL}, it will be ignored. (default = NULL)}

\item{to}{Character string specifiying the end day(inclusive).
It should be a \code{POSIXct} object or a \code{character} which can be casted to \code{POSIXct}.
If \code{NULL}, it will be ignored. (default = NULL)}

\item{split}{If \code{TRUE}, the result will be splitted by the station code.
Otherwise, not. (default = FALSE)}
}
\value{
A data frame holding the requested info if \code{split} is FALSE.
A list properly splitted by the station code if \code{split} is TRUE.
}
\description{
Take request and group specification, it returns corresponding data from the source DB.
In case of invalid request, it returns an NULL and prints out a warning message.
}
\examples{
# Get daily subway data of SNU and Shinchon station
# 228 and 240 are the station codes for SNU and Shinchon, respectively
data <- get.subway(c(228, 240), group = "day")
head(data)

# Get montly subway data of SNU and Shinchon station
# We want splitted data and want to mean over every month
data <- get.subway(c(228, 240), group = "month", FUN = mean, split = TRUE)
head(data$"228")
head(data$"240")

# Get weekly data of SNU station to 12/31/2019
data <- get.subway(228, group = "week", to = "2019-12-31")
tail(data)
}
\author{
Sanghyun Park, Daun Jeong, and Sehun Kim
}
