% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{model.select}
\alias{model.select}
\title{SARIMA model selection procedure}
\usage{
model.select(
  data,
  p,
  d,
  q,
  P,
  D,
  Q,
  period,
  ...,
  criterion = c("aic", "aicc", "bic"),
  plot = TRUE,
  verbose = TRUE,
  log = TRUE
)
}
\arguments{
\item{data}{Time series data for grid search.
The data should be either \code{ts} object or \code{numeric} vector.}

\item{p}{Candidates for AR order.}

\item{d}{Candidates for degree of differencing.}

\item{q}{Candidates for MA order.}

\item{P}{Candidates for seasonal AR order.}

\item{D}{Candidates for seasonal degree of differencing.}

\item{Q}{Candidates for seasonal MA order.}

\item{period}{Candidates for seasonal term period.
Note that long period needs quite long computation time.}

\item{...}{Extra arguments for SARIMA fitting.
Refer to \code{\link{Arima}}.}

\item{criterion}{Model selection criterion.
Should be one of \code{"aic"}, \code{"aicc"}, and \code{"bic"}.}

\item{plot}{If \code{TRUE}, the obtained values of criterion for each candidiate model will be plotted.
Otherwise, not. (default = TRUE)}

\item{verbose}{If \code{TRUE}, it prints out messages indicating the internal progress.
Otherwise, not.
Since model selection procedure can take a long time, we highly recommend to keep it \code{TRUE}. (default = TRUE)}
}
\value{
A list containing the following elements:\tabular{ll}{
  \code{min} \tab Achieved minimum of the criterion.\cr
  \tab\cr
  \code{select} \tab Selected model(\code{Arima} object).\cr
  \tab\cr
  \code{tune} \tab Data frame holding the grid search result. Ordered by the criterion values.\cr
  \tab\cr
  \code{models} \tab List of all models fitted. Ordered by the criterion values.\cr
  \tab\cr
  \code{plot} \tab \code{ggplot} object of the plot. (Returned only if \code{plot} is TRUE.)\cr
}
}
\description{
Take data and candidate parameters and run grid search.
For efficiency, it computes in parallel with heuristics on CPU load balancing.
For SARIMA model fitting, it uses \code{\link{Arima}} from \code{forecast}.
Sometimes, the fitting results in error for specific parameter combination (e.g. failiure in stationarity).
In such case, it just ignores such combination.
If \code{plot} is TRUE, it also plots the grid search result.
}
\examples{
# 228 is the station code for SNU
data <- get.subway(228)

# Model selection
# It would take some time
model.select(data$total, p = 1:3, d = 1:3, q = 1:3,
             P = 1:3, D = 1:3, Q = 1:3, period = c(1, 7),
             criterion = "aic")
}
\author{
Sanghyun Park, Daun Jeong, and Sehun Kim
}
