% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{normal.test}
\alias{normal.test}
\title{Normality check procedure for residuals}
\usage{
normal.test(fit, alpha = 0.05, plot = TRUE)
}
\arguments{
\item{fit}{Fitted ARIMA model(\code{ARIMA} object).}

\item{alpha}{The probability of type 1 error for normality tests. (default = 0.05)}

\item{plot}{If \code{TRUE}, the Q-Q plot of residuals and estimated density of residuals will be presented.
Otherwise, not. (default = TRUE)}
}
\value{
A list containing the following elements:\tabular{ll}{
  \code{shapiro.wilk} \tab The test result of Shapiro-Wilk test(\code{htest} objects).\cr
  \tab\cr
  \code{jarque.bera} \tab The test result of Jarque-Bera test(\code{htest} objects).\cr
  \tab\cr
  \code{komogorov.smirnov} \tab The test result of Kolmogorov-Smirnov test(\code{htest} objects).\cr
  \tab\cr
  \code{plot} \tab \code{ggplot} object of the plot. (Returned only if \code{plot} is TRUE.)\cr
}
}
\description{
Take fitted ARIMA model and perform normality checks.
It runs three well-known normality test, Shapiro-Wilk test, Jarque-Bera test, and Kolmogorov-Smirnov test
and prints out the result to the console.
If \code{plot} is TRUE, it also plots Q-Q plot of residuals and estimated density of residual for visual check.
}
\examples{
# 228 is the station code for SNU
data <- get.subway(228)

# Fit SARIMA model with total on and off passenger counts
SNU.total <- ts(data$total, frequency = 7)
SNU.fit <- auto.arima(SNU.total)

# Normality check
normal.test(SNU.fit)
}
\author{
Sanghyun Park, Daun Jeong, and Sehun Kim
}
