% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.pacf}
\alias{plot.pacf}
\title{Plot sample PACF of time series}
\usage{
plot.pacf(data, lag.max = NULL, conf.level = 0.95, name = NULL)
}
\arguments{
\item{data}{Time series data or list of them whose PACF is to be plotted.
The data should be either \code{ts} object or \code{numeric} vector.}

\item{lag.max}{Maximum length of lag to be shown.
If null, it uses default setting of \code{\link{acf}} function.}

\item{conf.level}{The confidence level for CI. (default = 0.95)}

\item{name}{Character vector of data names for facet titles.
If NULL, it uses the name of input data argument instead. (default = NULL)}
}
\value{
\code{ggplot} object of the plot
}
\description{
Take time series data and plot their sample PACF and CI.
}
\author{
Sanghyun Park, Daun Jeong, and Sehun Kim
}
