% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.raw}
\alias{plot.raw}
\title{Plot raw time series}
\usage{
plot.raw(data, date = NULL, name = NULL)
}
\arguments{
\item{data}{Time series data or list of them to be plotted.
The data should be either \code{ts} object or \code{numeric} vector.}

\item{date}{POSIXct vector or list of them for x axis.
If NULL, it uses integer index 1, 2, ... instead. (default = NULL)}

\item{name}{Character vector of data names for facet titles.
If NULL, it uses the name of input data argument instead. (default = NULL)}
}
\value{
\code{ggplot} object of the plot
}
\description{
Take time series data and plot raw traces.
}
\author{
Sanghyun Park, Daun Jeong, and Sehun Kim
}
