% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.residual}
\alias{plot.residual}
\title{Plot studentized residuals}
\usage{
plot.residual(fit, date = NULL, name = NULL)
}
\arguments{
\item{fit}{Fitted ARIMA model(\code{ARIMA} object) or list of them whose residuals to be plotted.}

\item{date}{POSIXct vector or list of them for x axis.
If NULL, it uses integer index 1, 2, ... instead. (default = NULL)}

\item{name}{Character vector of model names for facet titles.
If NULL, it uses the model name of input \code{ARIMA} object instead. (default = NULL)}
}
\value{
\code{ggplot} object of the plot
}
\description{
Take fitted ARIMA models and plot studentized residuals.
}
\author{
Sanghyun Park, Daun Jeong, and Sehun Kim
}
