% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.trend}
\alias{plot.trend}
\title{Plot trend of time series}
\usage{
plot.trend(
  data,
  method = c("loess", "ma", "lm", "gam", "glm"),
  ...,
  date = NULL,
  name = NULL
)
}
\arguments{
\item{data}{Time series data or list of them whose trend is to be plotted.
The data should be either \code{ts} object or \code{numeric} vector.}

\item{method}{Smoothing method.
Should be one of \code{"loess"}, \code{"ma"}, \code{"lm"}, \code{"gam"}, and \code{"glm"}. (default = "loess")}

\item{...}{Extra arguments for smoothing.
Refer to \code{\link{ma}} and \code{\link{geom_smooth}}.
In caes of MA, if \code{order} argument is not given, 10 will be as default value.}

\item{date}{POSIXct vector or list of them for x axis.
If NULL, it uses integer index 1, 2, ... instead. (default = NULL)}

\item{name}{Character vector of data names for facet titles.
If NULL, it uses the name of input data argument instead. (default = NULL)}
}
\value{
\code{ggplot} object of the plot
}
\description{
Take time series data and overlay their trend over raw trace.
}
\author{
Sanghyun Park, Daun Jeong, and Sehun Kim
}
