% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{psd}
\alias{psd}
\title{Power spectrum analaysis for time series}
\usage{
psd(data, plot = TRUE)
}
\arguments{
\item{data}{Time series data or list of them to be plotted.
The data should be either \code{ts} object or \code{numeric} vector.}

\item{plot}{If \code{TRUE}, the plot of estimated PSD will be presented.
Otherwise, not. (default = TRUE)}
}
\value{
A list containing the following elements:\tabular{ll}{
  \code{period} \tab Data frame holding suggested periods and corresponding PSD. Ordered by PSD.\cr
  \tab\cr
  \code{plot} \tab \code{ggplot} object of the plot. (Returned only if \code{plot} is TRUE.)\cr
}
}
\description{
Take time series data and perform power specturm analysis.
Through FFT, it lists up the most probable periods for seasonal term.
Up to 10 most likely suggestions will be printed out to the console based on estimated PSD(Power Spectrum Density).
}
\examples{
# 228 is the station code for SNU
data <- get.subway(228)
psd(data$total)
}
\author{
Sanghyun Park, Daun Jeong, and Sehun Kim
}
