% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{search.code}
\alias{search.code}
\title{Search for subway station code}
\usage{
search.code(name, exact = FALSE)
}
\arguments{
\item{name}{Keyword to be searched.
Accepts substring and regex.}

\item{exact}{If \code{TRUE}, it only finds exact match.
Otherwise, not. (default = FALSE)}
}
\description{
Take search keyword and search internal dictionary for station codes.
It supports regex and substring mathcing.
The search result will be printed out to the console.
}
\examples{
# What is the station code for SNU?
search.code("서울대")
}
\author{
Sanghyun Park, Daun Jeong, and Sehun Kim
}
